app/requirements.txt
streamlit==1.50.0
requests>=2.25.0
Pillow>=9.0.0